+++ src/test/java/lotto/view/util/NumberParserTest.java
@@ -1,4 +1,4 @@
-package lotto.controller.util;
+package lotto.view.util;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
